* commit 'a4da25fcfb00e5e766749b1d9af73e140877905c':
  Fix ramsize<<20 overflow #PSBM-43481

+++ Dispatcher/Dispatcher/CDspVmConfigManager.cpp
@@ -55,7 +55,7 @@ enum Defaults {DEFAULT_MEMGUARANTEE_PERCENTS = 40};
 ///////////////////////////////////////////////////////////////////////////////
 // struct MemGuarantee
 
-quint64 MemGuarantee::operator()(quint32 ramsize_)
+quint64 MemGuarantee::operator()(quint64 ramsize_)
 {
 	switch(m_type)
 	{
